Split Asian to Buddhist, Chinese, Hindu
