+++ holidays/calendars/__init__.py
@@ -12,9 +12,11 @@
 from datetime import date
 from datetime import timedelta as td
 
-from holidays.calendars.asian import _AsianLunisolar
+from holidays.calendars.buddhist import _BuddhistLunisolar
+from holidays.calendars.chinese import _ChineseLunisolar
 from holidays.calendars.custom import _CustomCalendar
 from holidays.calendars.hebrew import _HebrewLuniSolar
+from holidays.calendars.hindu import _HinduLunisolar
 from holidays.calendars.islamic import _IslamicLunar
 from holidays.calendars.thai import _ThaiLuniSolar
