+++ holidays/countries/china.py
@@ -13,11 +13,11 @@ from datetime import timedelta as td
 
 from holidays.constants import APR, OCT
 from holidays.holiday_base import HolidayBase
-from holidays.holiday_groups import AsianCalendarHolidays
+from holidays.holiday_groups import ChineseCalendarHolidays
 from holidays.holiday_groups import InternationalHolidays
 
 
-class China(HolidayBase, AsianCalendarHolidays, InternationalHolidays):
+class China(HolidayBase, ChineseCalendarHolidays, InternationalHolidays):
     """
     https://en.wikipedia.org/wiki/Public_holidays_in_China
     """
@@ -25,7 +25,7 @@ class China(HolidayBase, AsianCalendarHolidays, InternationalHolidays):
     country = "CN"
 
     def __init__(self, *args, **kwargs):
-        AsianCalendarHolidays.__init__(self)
+        ChineseCalendarHolidays.__init__(self)
         InternationalHolidays.__init__(self)
         super().__init__(*args, **kwargs)
