+++ holidays/countries/vietnam.py
@@ -13,11 +13,11 @@ from datetime import timedelta as td
 
 from holidays.constants import APR, SEP
 from holidays.holiday_base import HolidayBase
-from holidays.holiday_groups import AsianCalendarHolidays
+from holidays.holiday_groups import ChineseCalendarHolidays
 from holidays.holiday_groups import InternationalHolidays
 
 
-class Vietnam(HolidayBase, AsianCalendarHolidays, InternationalHolidays):
+class Vietnam(HolidayBase, ChineseCalendarHolidays, InternationalHolidays):
     """
     https://publicholidays.vn/
     http://vbpl.vn/TW/Pages/vbpqen-toanvan.aspx?ItemID=11013 Article.115
@@ -27,7 +27,7 @@ class Vietnam(HolidayBase, AsianCalendarHolidays, InternationalHolidays):
     country = "VN"
 
     def __init__(self, *args, **kwargs):
-        AsianCalendarHolidays.__init__(self)
+        ChineseCalendarHolidays.__init__(self)
         InternationalHolidays.__init__(self)
         super().__init__(*args, **kwargs)
